Move to a internal helper package
