+++ packages/integrations/vercel/package.json
@@ -51,6 +51,7 @@
   },
   "dependencies": {
     "@astrojs/webapi": "^2.1.1",
+    "@astrojs/internal-helpers": "^0.1.0",
     "@vercel/analytics": "^0.1.8",
     "@vercel/nft": "^0.22.1",
     "esbuild": "^0.17.12",
