+++ packages/internal-helpers/readme.md
@@ -1,0 +1,3 @@
+# @astrojs/internal-helpers
+
+These are internal helpers used by core Astro packages. This package does not follow semver and should not be used externally.
